+++ contracts/guard/src/SimpleVaultV0.sol
@@ -6,7 +6,6 @@
 pragma solidity ^0.8.0;
 
 import "@openzeppelin/access/Ownable.sol";
-import "@openzeppelin/token/ERC20/IERC20.sol";
 
 import "./GuardV0.sol";