+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Current
 
+- Add: ERC-4626 APY calculation, both live and historical
+- Add: Gnosis Safe guard support for ERC-4626 vaults using [TradingStrategyModuleV0](https://github.com/tradingstrategy-ai/web3-ethereum-defi/tree/master/contracts/safe-integration) Zodiac module
+- Add: [More ERC-4626 tutorials](https://web3-ethereum-defi.readthedocs.io/tutorials/index.html)
+- Update: Stablecoin token lists
+
+# 0.29.1
+
 - Add: ERC-4626 deposit/redeem
 - Add: ERC-4626 vault flow transaction analysis
 - Add: ERC-4626 share price redemption estimator
 - Fix: Upgrade Velvet vault swap API to their new version, old API no longer works
+- Bump dependencies, inc. Plotly
 
 # 0.29